// dunadev-frontend/src/api/dunadev/model/organiserProfile.ts
/**
 * DunaDev API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Full profile of the authenticated organiser.
 */
export interface OrganiserProfile { 
    id: number;
    name: string;
    description?: string | null;
    websiteUrl?: string | null;
    logoUrl?: string | null;
}