// dunadev-frontend/src/api/dunadev/model/organiserUpdateRequest.ts
/**
 * DunaDev API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrganiserUpdateRequest { 
    name: string;
    description?: string | null;
    websiteUrl?: string | null;
}